tests/requirements-test.txt
# TSUNAMI Test Dependencies
# Install with: pip install -r tests/requirements-test.txt

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0  # Parallel test execution

# Property-based testing
hypothesis>=6.92.0

# Mocking and fixtures
pytest-mock>=3.12.0
responses>=0.24.0  # Mock HTTP requests
aioresponses>=0.7.5  # Mock async HTTP requests

# Flask testing
pytest-flask>=1.3.0

# Code quality checks
flake8>=6.1.0
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0

# Security testing
bandit>=1.7.6  # Security linter
safety>=2.3.5  # Dependency vulnerability scanner

# Coverage reporting
coverage[toml]>=7.3.0

# Additional test utilities
faker>=22.0.0  # Generate fake data
freezegun>=1.2.2  # Mock datetime

# Optional: Browser testing (if needed)
# playwright>=1.40.0
# pytest-playwright>=0.4.3
